Unimportant menu stuff

+++ MoreThanWords2/Assets/Scripts/Managers/GameController.cs
@@ -12,15 +12,12 @@ using UnityEngine;
 public class GameController : MonoBehaviour
 {
     public static GameController gameController;
-
     [SerializeField] private GameObject mainMenu, optionsMainMenu, chaptersMainMenu;
-
-    private Button load, chapters;
-
+    [SerializeField] private Button load, chapters;
     private string saveLocation;
     //private string settingsLocation;
-
     [HideInInspector] public bool save;
+    public bool findObjects;
     //[HideInInspector] public bool noSettings;
 
     /*public float masterAudio;
@@ -32,17 +29,16 @@ public class GameController : MonoBehaviour
     public int resolutionY;*/
 
 
-    void Awake()
+    void Awake ()
     {
         saveLocation = Application.persistentDataPath + "/Level.dat";
         //settingsLocation = Application.persistentDataPath + "Settings.dat";
-
         save = true;
+        findObjects = true;
 
-        optionsMainMenu.SetActive(true);
-        mainMenu.SetActive(true);
-        optionsMainMenu.SetActive(false);
-        chaptersMainMenu.SetActive(false);
+        optionsMainMenu.SetActive (true);
+        mainMenu.SetActive (true);
+        chaptersMainMenu.SetActive (false);
 
         if (gameController == null)
         {
@@ -56,9 +52,7 @@ public class GameController : MonoBehaviour
 
         if (File.Exists(saveLocation) == false)
         {
-            load = GameObject.Find("ContinueBut").GetComponent<Button>();
             load.interactable = false;
-            chapters = GameObject.Find("ChaptersBut").GetComponent<Button>();
             chapters.interactable = false;
             //noSettings = true;
         }
@@ -80,6 +74,20 @@ public class GameController : MonoBehaviour
         if (Input.GetKeyDown(KeyCode.X) == true)
         {
             Delete();
+        }
+
+        if (findObjects == true && SceneManager.GetActiveScene().name == "Menu")
+        {
+            mainMenu = GameObject.Find ("MainMenuObj");
+            //chaptersMainMenu = GameObject.Find ("ChaptersMenuObj");
+            //chaptersMainMenu.SetActive (false);
+            optionsMainMenu = GameObject.Find ("OptionsMenu");
+            optionsMainMenu.SetActive (false);
+            load = GameObject.Find("ContinueBut").GetComponent<Button> ();
+            //chapters = GameObject.Find("ChaptersBut").GetComponent<Button> ();
+
+            findObjects = false;
+            print (findObjects);
         }
     }
 
@@ -166,12 +174,12 @@ public class GameController : MonoBehaviour
 
     public void Delete()
     {
-        if (File.Exists (saveLocation) == true)
+        if (SceneManager.GetActiveScene().name == "Menu" && File.Exists (saveLocation) == true)
         {
             File.Delete(saveLocation);
-            load = GameObject.Find("ContinueBut").GetComponent<Button>();
+            //load = GameObject.Find("ContinueBut").GetComponent<Button>();
             load.interactable = false;
-            chapters = GameObject.Find("ChaptersBut").GetComponent<Button>();
+            //chapters = GameObject.Find("ChaptersBut").GetComponent<Button>();
             chapters.interactable = false;
         }
     }
